# perf: implement filter query caching