--- a/src/filter.ts
+++ b/src/filter.ts
@@ -4,6 +4,8 @@ import { type Update } from "./types.ts";
 
 type FilterFunction<C extends Context, D extends C> = (ctx: C) => ctx is D;
 
+const filterQueryCache = new Map<string[], (ctx: Context) => boolean>();
+
 // === Obtain O(1) filter function from query
 /**
  * > This is an advanced function of grammY.
@@ -31,8 +33,13 @@ type FilterFunction<C extends Context, D extends C> = (ctx: C) => ctx is D;
 export function matchFilter<C extends Context, Q extends FilterQuery>(
     filter: Q | Q[],
 ): FilterFunction<C, Filter<C, Q>> {
-    const parsed = parse(filter);
-    const predicate = compile(parsed);
+    const queries = Array.isArray(filter) ? filter : [filter];
+    const predicate = filterQueryCache.get(queries) ?? (() => {
+        const parsed = parse(queries);
+        const pred = compile(parsed);
+        filterQueryCache.set(queries, pred);
+        return pred;
+    })();
     return (ctx: C): ctx is Filter<C, Q> => predicate(ctx);
 }